[android] updated wrapper for v4.3.0

+++ BlinkID/src/android/src/main/java/com/microblink/reactnative/recognizers/serialization/CroatiaCombinedRecognizerSerialization.java
@@ -40,7 +40,7 @@ public final class CroatiaCombinedRecognizerSerialization implements RecognizerS
         jsonResult.putBoolean("dateOfExpiryPermanent", result.isDateOfExpiryPermanent());
         jsonResult.putMap("dateOfIssue", SerializationUtils.serializeDate(result.getDateOfIssue()));
         jsonResult.putString("digitalSignature", SerializationUtils.encodeByteArrayToBase64(result.getDigitalSignature()));
-        jsonResult.putInt("digitalSignatureVersion", result.getDigitalSignatureVersion());
+        jsonResult.putInt("digitalSignatureVersion", (int)result.getDigitalSignatureVersion());
         jsonResult.putBoolean("documentBilingual", result.isDocumentBilingual());
         jsonResult.putBoolean("documentDataMatch", result.isDocumentDataMatch());
         jsonResult.putString("faceImage", SerializationUtils.encodeImageBase64(result.getFaceImage()));
